Initial Unity Project
I spent a bit of time developing a much better level generation script -
the old one is included too.

I'm not happy with the character controller quite yet due to the unity
physics, but we'll adjust as necessary.

Time spent 5 hours total on the level generation system, 1 hour on the
character controller.